requirements.txt
streamlit>=1.30.0
python-pptx>=0.6.21
Pillow>=10.0.0
lxml>=4.9.0
